stream: close() blocks until all written data is acked
in order to gracefully shut down a connection, the application must
not close it until all of its streams have finished their reliable
delivery

by default with tcp sockets, close() returns immediately but the kernel
may continue cleanup in the background, like (re)transmitting its packets
until everything gets acked

as a user space transport, we can't rely on the kernel to manage this
for us. the stream cleanup also depends on its associated connection,
which is managed by the application itself

default to the behavior of the socket option SO_LINGER, which causes
close() to block until all of its written data is acked. by waiting for
all of its streams to return from close(), the application can preserve
the lifetime of the connection long enough for a graceful shutdown

add async_close() to wait asynchronously

SO_LINGER also effects shutdown(), but the implementation is based on
the lsquic flag 'es_delay_onclose', which delays the final on_close()
callback as long as there's unacked data on the stream. so we'd be
unable to trigger the completion of shutdown(1) until another call
to shutdown(0) resulted in the stream to close itself

Signed-off-by: Casey Bodley <cbodley@gmail.com>